move to pulsar-common
