+++ pulsar-common/src/main/java/org/apache/pulsar/common/util/LatencyTracer.java
@@ -16,8 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-
-package org.apache.pulsar.utils;
+package org.apache.pulsar.common.util;
 
 import com.google.common.annotations.VisibleForTesting;
 import java.util.ArrayList;
